requirements.txt
# Core AI & LLM
google-generativeai>=0.8.0
google-ai-generativelanguage>=0.6.0

# Document Processing
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3

# Blockchain (Solana)
solana>=0.34.0
solders>=0.21.0
base58>=2.1.1

# Audio (ElevenLabs)
elevenlabs>=1.0.0
requests>=2.31.0

# Database & Storage
sqlalchemy>=2.0.0
python-dotenv>=1.0.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0

# Utilities
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
python-json-logger>=2.0.7
tenacity>=8.2.3

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
